requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiohttp-retry==2.9.1
aiosignal==1.3.2
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
asttokens==3.0.0
attrs==25.3.0
azure-core==1.33.0
azure-storage-blob==12.25.1
azure-storage-file-datalake==12.20.0
blinker==1.9.0
boto3==1.37.34
botocore==1.37.34
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
comm==0.2.2
contourpy==1.3.1
cryptography==44.0.2
cycler==0.12.1
databricks-agents==0.19.0
databricks-ai-bridge==0.4.1
databricks-connect==16.3.1
databricks-langchain==0.4.2
databricks-sdk==0.50.0
databricks-vectorsearch==0.55
dataclasses-json==0.6.7
debugpy==1.8.14
decorator==5.2.1
deprecated==1.2.18
deprecation==2.1.0
distro==1.9.0
docker==7.1.0
executing==2.2.0
fastapi==0.115.12
flask==3.1.0
fonttools==4.57.0
frozenlist==1.5.0
gitdb==4.0.12
gitpython==3.1.44
google-api-core==2.24.2
google-auth==2.39.0
google-cloud-core==2.4.3
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
grpcio==1.71.0
grpcio-status==1.71.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
importlib-metadata==8.6.1
ipykernel==6.29.5
ipython==9.1.0
ipython-pygments-lexers==1.1.1
isodate==0.7.2
itsdangerous==2.2.0
jedi==0.19.2
jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jupyter-client==8.6.3
jupyter-core==5.7.2
kiwisolver==1.4.8
langchain==0.3.23
langchain-community==0.3.21
langchain-core==0.3.52
langchain-experimental==0.3.4
langchain-openai==0.3.13
langchain-text-splitters==0.3.8
langgraph==0.3.4
langgraph-checkpoint==2.0.24
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.61
langsmith==0.3.31
mako==1.3.10
markdown==3.8
markdown-it-py==3.0.0
markupsafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mlflow==2.21.3
mlflow-skinny==2.21.3
multidict==6.4.3
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==1.26.4
openai==1.74.0
opentelemetry-api==1.32.1
opentelemetry-sdk==1.32.1
opentelemetry-semantic-conventions==0.53b1
orjson==3.10.16
ormsgpack==1.9.1
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
pip==25.0.1
platformdirs==4.3.7
prompt-toolkit==3.0.51
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
ptyprocess==0.7.0
pure-eval==0.2.3
py4j==0.10.9.7
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.22
pydantic==2.11.3
pydantic-core==2.33.1
pydantic-settings==2.8.1
pygments==2.19.1
pymupdf==1.25.5
pyparsing==3.2.3
pypdf==5.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pyyaml==6.0.2
pyzmq==26.4.0
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
rsa==4.9
ruff==0.11.5
s3transfer==0.11.4
scikit-learn==1.6.1
scipy==1.15.2
setuptools==78.1.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sqlalchemy==2.0.40
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.46.2
tabulate==0.9.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing-extensions==4.13.2
typing-inspect==0.9.0
typing-inspection==0.4.0
tzdata==2025.2
unitycatalog-ai==0.3.0
unitycatalog-client==0.2.1
unitycatalog-langchain==0.2.0
urllib3==2.4.0
uvicorn==0.34.1
wcwidth==0.2.13
werkzeug==3.1.3
wrapt==1.17.2
yarl==1.19.0
zipp==3.21.0
zstandard==0.23.0
